requirements-bak.txt
absl-py==2.1.0
accelerate==1.0.1
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
albumentations==0.5.2
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.5.2
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.3.0
backcall==0.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.38.1
bleach==6.1.0
blinker==1.8.2
blis==0.7.11
braceexpand==0.1.7
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==2.1.1
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpathlib==0.20.0
coloredlogs==15.0.1
colorlog==6.9.0
confection==0.1.5
contexttimer==0.3.3
contourpy==1.1.1
controlnet-aux==0.0.9
cycler==0.12.1
cymem==2.0.10
decorator==5.1.1
decord==0.6.0
defusedxml==0.7.1
diff_gaussian_rasterization @ file:///media/ps/disk2/code/Edit3D/edit_3dgs/GaussianEditor/gaussiansplatting/submodules/diff-gaussian-rasterization
diffusers==0.19.3
distlib==0.3.9
docker-pycreds==0.4.0
docstring_parser==0.16
easydict==1.13
einops==0.8.0
embreex==2.17.7.post5
envlight @ git+https://github.com/ashawkey/envlight.git@05b5851e854429d72ecaf5b206ed64ce55fae677
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.1.0
fairscale==0.4.4
fastapi==0.115.6
ffmpy==0.5.0
filelock==3.13.1
flatbuffers==24.12.23
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.2.0
ftfy==6.2.3
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.37.0
google-auth-oauthlib==1.0.0
gradio==3.50.2
gradio_client==0.6.1
groundingdino @ git+https://github.com/IDEA-Research/GroundingDINO.git@856dde20aee659246248e20734ef9ba5214f5e44
grpcio==1.68.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.23.3
humanfriendly==10.0
identify==2.6.1
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
imgaug==0.4.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
iopath==0.1.10
ipython==8.12.3
jax==0.4.13
jaxlib==0.4.13
jaxtyping==0.2.19
jedi==0.19.2
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kaggle==1.6.17
kiui @ git+https://github.com/ashawkey/kiuikit@73e34203fd3bd4b5979dc3a38057dff117b39521
kiwisolver==1.4.7
kornia==0.7.0
langcodes==3.4.1
language_data==1.3.0
lazy_loader==0.4
libigl==2.5.1
lightning-utilities==0.11.9
llvmlite==0.41.1
lxml==5.3.0
manifold3d==3.0.1
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.11
mediapy==1.2.2
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.18.6
multidict==6.1.0
murmurhash==1.0.11
narwhals==1.20.1
networkx==3.0
ninja==1.11.1.3
nodeenv==1.9.1
numba==0.58.1
numpy==1.24.1
nvdiffrast @ git+https://github.com/NVlabs/nvdiffrast.git@729261dc64c4241ea36efda84fbf532cc8b425b8
oauthlib==3.2.2
objprint==0.2.3
omegaconf==2.3.0
onnxruntime==1.16.3
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
opt_einsum==3.4.0
orjson==3.10.13
packaging==24.2
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
plotly==5.24.1
plyfile==1.0.3
pooch==1.8.2
portalocker==3.0.0
pre-commit==3.5.0
preshed==3.0.9
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==3.20.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocoevalcap==1.2
pycocotools==2.0.7
pycollada==0.8
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyliblzfse==0.4.1
PyMatting==1.1.13
PyMCubes==0.1.4
pyparsing==3.1.4
pySciTools==0.1.2
pysdf==0.1.9
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-magic==0.4.27
python-multipart==0.0.20
python-slugify==8.0.4
pytorch-lightning==2.4.0
pytz==2024.2
PyWavelets==1.4.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
rembg==2.0.61
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.20.1
rsa==4.9
Rtree==1.3.0
ruff==0.8.4
safetensors==0.4.5
salesforce-lavis @ git+https://github.com/salesforce/LAVIS.git@a557de5df3a59bde8d9a8cde3af89cb2db41bd31
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.0
segment-anything==1.0
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.19.2
setproctitle==1.3.4
shapely==2.0.6
shellingham==1.5.4
shtab==1.7.1
simple_knn @ file:///media/ps/disk2/code/Edit3D/edit_3dgs/GaussianEditor/gaussiansplatting/submodules/simple-knn
six==1.17.0
smart-open==7.1.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.41.3
streamlit==1.40.1
supervision==0.25.1
svg.path==6.3
sympy==1.13.1
taming-transformers-rom1504==0.0.6
tenacity==9.0.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
text-unidecode==1.3
thinc==8.2.4
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==0.4.12
tinycudann @ git+https://github.com/NVlabs/tiny-cuda-nn/@c91138bcd4c6877c8d5e60e483c0581aafc70cce#subdirectory=bindings/torch
tokenizers==0.13.3
toml==0.10.2
tomli==2.2.1
tomlkit==0.12.0
torch==2.1.1+cu118
torch_efficient_distloss==0.1.3
torchaudio==2.1.1+cu118
torchmetrics==1.5.2
torchvision==0.16.1+cu118
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.25.0
trimesh==4.5.3
triton==2.1.0
typeguard==4.4.0
typer==0.15.1
typing_extensions==4.12.2
tyro==0.9.5
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.33.0
varname==0.13.5
virtualenv==20.28.0
-e git+https://github.com/heheyas/viser@9704a5157a599c4e83242b71b01fbaa26e506b10#egg=viser
wandb==0.19.1
wasabi==1.1.3
watchdog==4.0.2
wcwidth==0.2.13
weasel==0.4.1
webdataset==0.2.100
webencodings==0.5.1
websockets==11.0.3
Werkzeug==3.0.6
wrapt==1.17.0
xatlas==0.0.9
xformers==0.0.23+cu118
xxhash==3.5.0
yapf==0.43.0
yarl==1.15.2
yourdfpy==0.0.56
zipp==3.20.2
